% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{all_types}
\alias{all_types}
\title{Summary statistics for all types in a Multitype Point Pattern}
\usage{
all_types(points, which.marks = NULL, ...)
}
\arguments{
\item{points}{multitype \code{\link{ppp}} object}

\item{which.marks}{string specifying column of the \code{\link{ppp}}
\code{marks} data.frame to use as marks for cross-
type analysis}

\item{...}{argument passed to \code{\link{alltypes}}}
}
\value{
Object of class "fv" - see \code{\link{fv.object}}
}
\description{
Wrapper for \code{\link{alltypes}} that that allows for a
subset of the input \code{\link{ppp}} \code{marks} to be used,
as specified by the parameter \code{which.marks}.
}
\seealso{
\code{\link{alltypes}}
}

