% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_parallel}
\alias{check_parallel}
\title{Simple function to establish whether things can be run in parallel using
\code{\link{foreach}}}
\usage{
check_parallel(verbose = TRUE)
}
\arguments{
\item{verbose}{specify whether to show output messages.}
}
\value{
boolean specifying whether a parallel backend has been setup.
}
\description{
Checks whether a parallel backend is setup for the \code{\link{foreach}}
package.
}

