% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{clear_border}
\alias{clear_border}
\title{Clear connected regions touching edges}
\usage{
clear_border(image, win = NULL, r = image$xstep * 2, output_mask = FALSE)
}
\arguments{
\item{image}{binary image (instance of \code{\link{im}} object)}

\item{win}{Optional.
Window/mask that defines edges (instance of \code{\link{owin}} object)}

\item{r}{Optional.
Radius of morphological erosion operation to use to define the edge. This
will determine the 'thickness' of the edge. Defaults to 2px.}
}
\value{
binary image (\code{\link{im}} object) excluding any connected
     components that touch the edge.
}
\description{
Clear connect regions of a binary image that touch either the edges of the
image or edges in specified binary mask (or window).
}

