% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{co_cluster_l_cross}
\alias{co_cluster_l_cross}
\title{Co-cluster analysis using self- and cross-type local L-function analysis}
\usage{
co_cluster_l_cross(points, rvalue, which.marks = NULL,
  mark.original = FALSE, ...)
}
\arguments{
\item{points}{A point pattern (object of class \code{ppp})}

\item{rvalue}{A single value of the distance argument r at which
the L-function should be computed.}

\item{which.marks}{Specifies the column of marks to use as types. Must be
specified is the PPP has multiple mark categories.}

\item{mark.original}{Logical value specifying whether the original point
pattern should be marked with the output and returned.}

\item{...}{Extra arguments are passed to the \code{local_l}
and \code{local_l_cross} functions.}
}
\value{
Self- and cross-type local L-functions values for each point are
 returned as either:
   \itemize{
     \item data.frame (if \code{mark.original} is \code{FALSE})
     \item marks in the original points (\code{ppp}) object passed as
           as argument (if \code{mark.original} is \code{TRUE})
   }
}
\description{
Performs self- and cross-type local L-function analysis at a specified
radius on a multi-type PPP object. The column of marks to use for types
is specified using the \code{which.marks}. Users have the option to return
either a data.frame containing 2 columns with the self- and cross-type
local L-function values or the original PPP where the self- and cross-type
local L-function value are added to the marks.
}

