% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{co_cluster_quadrant}
\alias{co_cluster_quadrant}
\title{Quadrant analysis of self- vs. cross local L-function values}
\usage{
co_cluster_quadrant(points, threshold, ll, llcross, mark.original = TRUE)
}
\arguments{
\item{points}{point pattern (multi-type \code{ppp} object)}

\item{threshold}{threshold that splits x and y axis of the quadrant plot}

\item{ll}{Column name of the self-type local L values}

\item{llcross}{Column names of the cross-type local L values}

\item{mark.original}{Logical value indicating whether labels for each point should
be added back to the original point pattern as marks or return as a vector.}
}
\value{
vector or point pattern depending on \code{mark.original}
}
\description{
Divides a relationship between self- and cross-type local
L-function values into quadrants based on a sepcified threshold. Determines
the quadrant number in which each point in a multi-type point pattern lies,
where quadrants are labelled:
\enumerate{
   \item lower left
   \item lower right
   \item upper left
   \item upper right
 }
}

