% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{exchange_xy_for_marks}
\alias{exchange_xy_for_marks}
\title{Exchange \code{x} and \code{y} coords for others stored as marks}
\usage{
exchange_xy_for_marks(points, x_name = "x", y_name = "y")
}
\arguments{
\item{points}{ppp object containing marks}

\item{x_name}{name of the marks column to be used as the new \code{x} coords}

\item{y_name}{name of the marks column to be used as the new \code{y} coords}
}
\value{
\code{\link{ppp}} where x and y coords have been exchanged for those
  those specified as marks.
}
\description{
Utility function that allows the \code{x} and/or \code{y} coords
of a PPP object to be swapped with another set of coords stored as marks.
The previous \code{x} and \code{y} coords are stored in the marks under
names \code{orig_x} and \code{orig_y} respectively.
}

