% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{fortify.im}
\alias{fortify.im}
\title{Fortify method for \code{\link{im}} objects}
\usage{
\method{fortify}{im}(model, data, ...)
}
\arguments{
\item{model}{im object}

\item{data}{not used by this method}

\item{...}{ignored by this method}
}
\value{
data.frame with 3 columns: \code{x}, \code{y} and \code{value}
 where \code{x} & \code{y} are the pixel coordinates, \code{value} is the
 intensity value at that pixel coordinate.
}
\description{
This method extracts coordinate and intensity data from a
\code{\link{im}} to a data frame, so that it can be plotted
with ggplot.
}
\seealso{
\code{\link{im}}
}

