% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{fortify.owin}
\alias{fortify.owin}
\title{Fortify method for unpacking \code{\link{owin}} objects.}
\usage{
\method{fortify}{owin}(model, data, ...)
}
\arguments{
\item{model}{owin object}

\item{data}{not used by this method}

\item{...}{ignored}
}
\value{
data.frame containing windows information. Information in the
 data.frame depends on the type of the \code{\link{owin}}:
 \itemize{
   \item \emph{rectangular windows}: data.frame with a single row and
     4 columns that contain xmin, xmax, ymin, ymax extents of the recangle.
   \item \emph{polygonal window}: data.frame with three columns: \code{id},
     \code{x}, \code{y} where \code{id} represents a unique id for each
     polygon when there are multiple, and \code{x} & \code{y} represent the
     x and y coordinates of each vertex in the polygon.
   \item \emph{mask window}: data.frame with three columns: \code{x},
     \code{y} and \code{value} where \code{x} & \code{y} are pixel
     coordinates, and \code{value} is a logical value denoting whether the
     pixel is part of the window.
 }
}
\description{
This method extracts relevant coordinates for plotting a
\link{\code{owin}} object to allow it to be plotted with ggplot.
}
\seealso{
\code{\link{owin}}
}

