% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{fortify.ppp}
\alias{fortify.ppp}
\title{Fortify method for \code{\link{ppp}} objects}
\usage{
\method{fortify}{ppp}(model, data, ...)
}
\arguments{
\item{model}{ppp object}

\item{data}{not used by this method}

\item{...}{ignored by this method}
}
\value{
data.frame containing at least two columns: \code{x} and \code{y}
 which represent the \code{x} and \code{y} coordinates of each point in a
 \code{\link{ppp}} object. Returned data.frame can also contain
 multiple other columns depending of the type of marks associated with the
 \code{\link{ppp}} object (see \code{\link{ppp}}). If mark type is:
 \itemize{
   \item \emph{vector}: a single column with heading \code{marks} will be added.
   \item \emph{data.frame}: the marks data.frame with be column bound to the
     output data.frame, where the column headings from the marks data.frame
     are preserved.
 }
}
\description{
This method extracts coordinate and marks data from
\code{\link{ppp}}objects to a data.frame, so that it can be
plotted with ggplot.
}
\seealso{
\code{\link{ppp}}
}

