% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vis.R
\name{fortify.psp}
\alias{fortify.psp}
\title{Fortify method for \code{\link{psp}} objects}
\usage{
\method{fortify}{psp}(model, data, ...)
}
\arguments{
\item{model}{psp object}

\item{data}{not used by this method}

\item{...}{ignored by this method}
}
\value{
data.frame containing at least four columns: \code{x0}, \code{x1},
 \code{y0} and \code{y1} with represent the coordinates of a line segement.
 Returned data.frame can also contain multiple other columns depending of
 the type of marks associated with the \code{\link{ppp}} object
 (see \code{\link{ppp}}). If mark type is:
 \itemize{
   \item \emph{vector}: a single column with heading \code{marks} will be added.
   \item \emph{data.frame}: the marks data.frame with be column bound to the
     output data.frame, where the column headings from the marks data.frame
     are preserved.
 }
}
\description{
This method extracts coordinate and marks data from
\code{\link{psp}} objects to a data.frame, so that it can be
plotted with ggplot.
}
\seealso{
\code{\link{psp}}
}

