% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{interpolate_surface}
\alias{interpolate_surface}
\title{Interpolate a surface over a spatial point pattern}
\usage{
interpolate_surface(matlab, data, xsize = 10, ysize = 10, method = "v4")
}
\arguments{
\item{matlab}{\code{Matlab} instance}

\item{data}{marked point pattern (\code{\link{ppp}} object) with
with a single vector of marks to be interpolated as a surface}

\item{xsize}{x pixel size of output image}

\item{ysize}{y pixel size of output image}

\item{method}{interpolation method for matlab gridata function, choices are
"nearest", "linear", "natural", "cubic" and "v4" (default).}
}
\value{
\code{\link{im}} of interpolated surface with pixel sizes
   specified by \code{xsize} and \code{ysize}.
}
\description{
Grid interpolation to create a surface map over a spatial point pattern
where the marks of the point pattern are interpolated between the points.
Function currently uses matlab \code{griddata} function for interpolation
using the \code{R.matlab} package.
}

