% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/localk_multi.R
\name{local_k_cross_inhom}
\alias{local_k_cross_inhom}
\alias{local_l_cross_inhom}
\title{Inhomogeneous multitype (cross-type) neighbourhood density function}
\usage{
local_k_cross_inhom(X, I, J, lambdaI = NULL, lambdaJ = NULL, ...,
  wantL = FALSE, correction = "Ripley", verbose = TRUE, rvalue = NULL,
  lambdaIJ = NULL, sigma = NULL, varcov = NULL)
}
\arguments{
\item{X}{Multitype point pattern (\code{\link{ppp}} object). It must be a multitype
point pattern (or marked point pattern).}

\item{lambdaI}{Optional. Values of the estimated intensity of the sub-process
of points of type \code{i}. Either a pixel image (object of class \code{"im"}),
a numeric vector containing the intensity values at each of the type
\code{i} points in \code{X}, or a \code{function(x,y)} which can be
evaluated to give the intensity value at any location.}

\item{lambdaJ}{Optional. Values of the estimated intensity of the sub-process
of points of type \code{j}. Either a pixel image (object of class \code{"im"}),
a numeric vector containing the intensity values at each of the type
\code{j} points in \code{X}, or a \code{function(x,y)} which can be
evaluated to give the intensity value at any location.}

\item{...}{Ignored.}

\item{correction}{String sprecifying the edge correction to be applied.
Option are \code{"none"}, \code{"translate"}, \code{"translation"},
\code{"Ripley"}, \code{"isotropic"} or \code{"best"}. Only one correction
may be given.}

\item{verbose}{Logical flag indicating whether to print progress reports
during the calculation.}

\item{rvalue}{Optional. A \emph{single} value of the distance argument \eqn{r}
at which the function L or K should be computed.}

\item{i}{The type (mark value) of the points in X from which distances are
measured. A character string (or something that will be converted to a
character string). Defaults to the first level of \code{\link{marks}(X)}.}

\item{j}{The type (marks value) of the points in X to which distances are
measured. A character string (or something that will be converted to a
character string'). Defaults to the second level of \code{\link{marks}(X)}.}
}
\value{
If \code{rvalue} is given, the result is a numeric vector of equal
     length to the number of points in X_i.

     If the code{rvalue} is absent, the result is an object of class \code{"fv"},
     see \code{\link{fv.object}}, which can be plotted directly using
     \code{\link{plot.fv}}. Essentially a data frame containing columns:
     \item{r}{the vector of values of the argument \eqn{r}
     at which the function \eqn{K} has been  estimated }
     \item{theo}{the theoretical value \eqn{K(r) = \pi r^2}{K(r) = pi * r^2}
     or \eqn{L(r)=r} for a stationary Poisson process
     }
     together with columns containing the values of the
     neighbourhood density function for each point in the pattern.
     Column \code{i} corresponds to the \code{i}th point.
     The last two columns contain the \code{r} and \code{theo} values.
}
\description{
Computes spatially-weighted versions of the local \eqn{K}-function or
\eqn{L}-function, defined by Getis and Franklin (1987).
}
\details{
The command \code{local_l_cross_inhom} computes the \emph{neighbourhood density function},
a local version of the \eqn{L}-function (Besag's transformation of Ripley's
\eqn{K}-function) proposed by Getis and Franklin (1987), for 2 types in a multitype
spatial point pattern. The command \code{local_k_cross} computes the corresponding
local analogue of the cross-type K-function.

Given a multitype spatial point pattern \code{X} with types \code{i} and \code{j},
the neighbourhood density function
\eqn{L_{ij}(r)}{L[ij](r)} associated with the \eqn{i}th point
in \code{X} is computed by
\deqn{
  L_{ij}(r) = \sqrt{\frac a {(n_j) \pi} \sum_j e_{ij}}
}{
  L[ij](r) = sqrt( (a/((n[j])* pi)) * sum[j] e[i,j])
}
where the sum is over all points \eqn{j}{j} that lie
within a distance \eqn{r} of the \eqn{i}th point,
\eqn{\lambda_j}{\lambda[j]} is the estimated intensity of type \eqn{j} of the
point pattern at the point \eqn{j}, and \eqn{e_{ij}}{e[i,j]} is an edge correction
term (as described in \code{\link{Kest}}).
The value of \eqn{L_{ij}(r)}{L[ij](r)} can also be interpreted as one
of the summands that contributes to the global estimate of the cross-type \eqn{L}-
function.

By default, the function \eqn{L_{ij}(r)}{L[ij](r)} or
\eqn{K_{ij}(r)}{K[ij](r)} is computed for a range of \eqn{r} values
for each point \eqn{i}. The results are stored as a function value
table (object of class \code{"fv"}) with a column of the table
containing the function estimates for each point \eqn{i} of the pattern
\code{X}.

Alternatively, if the argument \code{rvalue} is given, and it is a
single number, then the function will only be computed for this value
of \eqn{r}, and the results will be returned as a numeric vector,
with one entry of the vector for each point \eqn{i} of the pattern \code{X}.

Computation can be done in parallel by registering a parallel backend for
the \code{\link{foreach}} package.
}
\references{
Getis, A. and Franklin, J. (1987)
     Second-order neighbourhood analysis of mapped point patterns.
     \emph{Ecology} \bold{68}, 473--477.
}
\seealso{
\code{\link{localLinhom}}
         \code{\link{localKinhom}}
         \code{\link{Lcross.inhom}}
         \code{\link{Kcross.inhom}}
}

