% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localk.R
\name{local_k_inhom}
\alias{local_k_inhom}
\alias{local_l_inhom}
\title{Inhomogeneous Neighbourhood density function}
\usage{
local_k_inhom(X, lambda = NULL, ..., correction = "Ripley",
  verbose = TRUE, rvalue = NULL, sigma = NULL, varcov = NULL)
}
\arguments{
\item{X}{A point pattern (object of class \code{"ppp"}).}

\item{lambda}{Optional.
Values of the estimated intensity function.
Either a vector giving the intensity values
at the points of the pattern \code{X},
a pixel image (object of class \code{"im"}) giving the
intensity values at all locations, a fitted point process model
(object of class \code{"ppm"}) or a \code{function(x,y)} which
can be evaluated to give the intensity value at any location.}

\item{...}{ignored.}

\item{correction}{tring specifying the edge correction to be applied.
Options are \code{"none"}, \code{"translate"}, \code{"translation"},
\code{"Ripley"},
\code{"isotropic"} or \code{"best"}.
Only one correction may be specified.}

\item{verbose}{Logical flag indicating whether to print progress
reports during the calculation.}

\item{rvalue}{Optional. A \emph{single} value of the distance argument
\eqn{r} at which the function L or K should be computed.}

\item{sigma, }{varcov
Optional arguments passed to \code{\link{density.ppp}} to control
the kernel smoothing procedure for estimating \code{lambda},
if \code{lambda} is missing.}
}
\value{
If \code{rvalue} is given, the result is a numeric vector
 of length equal to the number of points in the point pattern.

 If \code{rvalue} is absent, the result is
 an object of class \code{"fv"}, see \code{\link{fv.object}},
 which can be plotted directly using \code{\link{plot.fv}}.
 Essentially a data frame containing columns
 \item{r}{the vector of values of the argument \eqn{r}
   at which the function \eqn{K} has been  estimated
 }
 \item{theo}{the theoretical value \eqn{K(r) = \pi r^2}{K(r) = pi * r^2}
  or \eqn{L(r)=r} for a stationary Poisson process
 }
 together with columns containing the values of the
 neighbourhood density function for each point in the pattern.
 Column \code{i} corresponds to the \code{i}th point.
 The last two columns contain the \code{r} and \code{theo} values.
}
\description{
Computes spatially-weighted versions of the local \eqn{K}-function or \eqn{L}-function.
Note: Equivalent to \code{\link{localKinhom}} and \code{\link{localLinhom}}
except that they can take advantage of parallel computation \eqn{K}-function
or \eqn{L}-function via \code{\link{foreach}} package.
}
\details{
The functions \code{local_k_inhom} and \code{local_l_inhom}
are inhomogeneous or weighted versions of the
neighbourhood density function implemented in
\code{\link{local_k}} and \code{\link{local_l}}.

Given a spatial point pattern \code{X}, the inhomogeneous neighbourhood
density function \eqn{L_i(r)}{L[i](r)} associated with the \eqn{i}th point
in \code{X} is computed by
\deqn{
  L_i(r) = \sqrt{\frac 1 \pi \sum_j \frac{e_{ij}}{\lambda_j}}
}{
  L[i](r) = sqrt( (1/pi) * sum[j] e[i,j]/lambda[j])
}
where the sum is over all points \eqn{j \neq i}{j != i} that lie
within a distance \eqn{r} of the \eqn{i}th point,
\eqn{\lambda_j}{\lambda[j]} is the estimated intensity of the
point pattern at the point \eqn{j}, and \eqn{e_{ij}}{e[i,j]} is an edge correction
term (as described in \code{\link{Kest}}).
The value of \eqn{L_i(r)}{L[i](r)} can also be interpreted as one
of the summands that contributes to the global estimate of the inhomogeneous L
function (see \code{\link{Linhom}}).

By default, the function \eqn{L_i(r)}{L[i](r)} or
\eqn{K_i(r)}{K[i](r)} is computed for a range of \eqn{r} values
for each point \eqn{i}. The results are stored as a function value
table (object of class \code{"fv"}) with a column of the table
containing the function estimates for each point of the pattern
\code{X}.

Alternatively, if the argument \code{rvalue} is given, and it is a
single number, then the function will only be computed for this value
of \eqn{r}, and the results will be returned as a numeric vector,
with one entry of the vector for each point of the pattern \code{X}.

Computation can be done in parallel by registering a parallel backend for
the \code{\link{foreach}} package.
}
\references{
Getis, A. and Franklin, J. (1987)
     Second-order neighbourhood analysis of mapped point patterns.
     \emph{Ecology} \bold{68}, 473--477.
}
\seealso{
\code{\link{Kinhom}}
         \code{\link{Linhom}}
         \code{\link{localKinhom}}
         \code{\link{localLinhom}}
         \code{\link{local_k}}
         \code{\link{local_l}}
}

