% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{local_max}
\alias{local_max}
\title{Feature detection functions.
Determine local max from a subset of a planar point pattern}
\usage{
local_max(df, mask)
}
\description{
Feature detection functions.
Determine local max from a subset of a planar point pattern
}

