% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{multi_local_l}
\alias{multi_local_l}
\title{Local L-function analysis}
\usage{
multi_local_l(points, which.marks = spatstat::marks(points), rvalue = NULL,
  mark.original = FALSE, ...)
}
\arguments{
\item{points}{multitype \code{\link{ppp}} object}

\item{which.marks}{string specifies the column of the \code{\link{ppp}}
\code{marks} data.frame that denotes points types.}

\item{rvalue}{radius of to determine L-function at.}

\item{mark.original}{if rvalue is specified, local L-function values can be
appended to the original \code{\link{ppp}} object as a new column
in \code{marks}.}

\item{...}{argument passed to \code{\link{local_l}}}
}
\value{
If \code{rvalue} and \code{mark.original} are specified,
   a vector of L(r) values is appended to the original \code{\link{ppp}}
   marks. If mark.original is not specified then the L(r) vector is simply returned.
   If \code{rvalue} is not specified, \code{\link{fv}} object is returned.
}
\description{
Wrapper for \code{\link{local_l}} that allows local L-function to
be computed on each type of a multitype \code{\link{ppp}} object
where type is specidied by the \code{marks} column denoted by the parameter
\code{which.marks}.
}
\details{
If an r-value is specified, \code{\link{local_l}} values can be appended as
marks on the original \code{\link{ppp}} object. Otherwise, a
\code{\link{fv}} object is returned.
}

