% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster.R
\name{multi_local_lcross}
\alias{multi_local_lcross}
\title{Cross-type Local L-function analysis}
\usage{
multi_local_lcross(points, which.marks = NULL, rvalue = NULL,
  mark.original = FALSE, ...)
}
\arguments{
\item{points}{multitype \code{\link{ppp}} object}

\item{which.marks}{string specifies the column of the \code{\link{ppp}}
\code{marks} data.frame that denotes points types.}

\item{rvalue}{radius of to determine L-function at.}

\item{mark.original}{if rvalue is specified, local L-function values can be
appended to the original \code{\link{ppp}} object as a new column
in \code{marks}.}

\item{...}{argument passed to \code{\link{local_l}}}
}
\value{
If \code{rvalue} and \code{mark.original} are specified,
   a vector of L(r)cross values is appended to the original \code{\link{ppp}}
   marks. If mark.original is not specified then the L(r)cross vector is simply returned.
   If \code{rvalue} is not specified, \code{\link{fv}} object is returned.
}
\description{
Wrapper for \code{\link{local_l_cross}} that allows cross-type
local L-function to be computed for specified multitype
\code{\link{ppp}} object using \code{marks} specified by the
parameter \code{which.marks}.
}
\details{
If an r-value is specified, localL values can be appended as marks on the
original \code{\link{ppp}} object.
}
\seealso{
\code{\link{local_k_cross}}
         \code{\link{local_l_cross}}
}

