% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{peak_local_max}
\alias{peak_local_max}
\title{Feature detection functions.
Peak local max}
\usage{
peak_local_max(points, min_distance, threshold = 50, which.marks = NULL)
}
\arguments{
\item{points}{Marked planar point pattern (ppp)}

\item{min_distance}{Min distance around a local max in which other maxima
cannot occur}

\item{threshold}{Minimal value that maxima can have (default = 50).}

\item{which.marks}{If ppp is marked with multiple types, this specifies
which marks to calculated peaks from.}
}
\value{
ppp containing only the detected maxima.
}
\description{
Determine local maxima of marked point pattern. Indentifies
local maxima of marks and returns the point pattern with only
those marks.
}

