% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{peak_r}
\alias{peak_r}
\title{Calculate r value at L(r) - r peaks}
\usage{
peak_r(lest, variable)
}
\arguments{
\item{lest}{fv output from \code{\link{Lest}} function}

\item{variable}{names of the curve for which to calculate peak r}
}
\value{
num peak r-value
}
\description{
Calculate r value at L(r) - r peaks
}
\seealso{
\code{\link{Lest}}
}

