% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.moleculelist_nstorm}
\alias{read.moleculelist_nstorm}
\title{Nikon N-STORM molecule list reader}
\usage{
read.moleculelist_nstorm(filepath, nonspecific = TRUE)
}
\arguments{
\item{filepath}{string denoting the path to the file to be read.}

\item{nonspecific}{include 'Non Specific Activation' data (default = TRUE)}
}
\value{
\code{\link{ppp}} object containing the coordinates for
         single molecule localisations and other parameters/columns in
         original csv file attached as marks.
}
\description{
Reader function for comma separated molecule list output of the
NIS N-STORM module.
}

