% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.moleculelist_thunderstorm}
\alias{read.moleculelist_thunderstorm}
\title{Thunderstorm molecule list reader}
\usage{
read.moleculelist_thunderstorm(filepath)
}
\arguments{
\item{filepath}{string denoting the path to the file to be read.}
}
\value{
\code{\link{ppp}} object containing the coordinates for
         single molecule localisations and other parameters/columns in
         original csv file attached as marks.
}
\description{
Reader function for comma separated molecule list output of the
Thunderstorm plugin for ImageJ/Fiji.
}

