% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.moleculelists}
\alias{read.moleculelists}
\title{Function to read multiple datasets using a specific reader}
\usage{
read.moleculelists(file_paths, rdr = read.moleculelist_rapidstorm, ...)
}
\arguments{
\item{file_paths}{list of file_paths to read}

\item{rdr}{reader function that takes a file path as input and returns
a ppp object.}

\item{...}{extra arguments passed to \code{rdr} function.}
}
\value{
\code{list(\link{ppp})}
}
\description{
Function to read multiple datasets using a specific reader
}

