% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.polygon_roi}
\alias{read.polygon_roi}
\title{Read a polygon owin from a tab-delimited txt file}
\usage{
read.polygon_roi(file_path, xsize = 1, ysize = 1)
}
\arguments{
\item{file_path}{path to txt file}

\item{xsize}{x pixel size}

\item{ysize}{y pixel size}
}
\value{
\code{\link{owin}}
}
\description{
Reads x and y coordinates from a tab delimited files where each XY coord is
on a separate row and X and Y coords are in columns 1 and 2, respectively.
}

