% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.rois}
\alias{read.rois}
\title{Function to read multiple windows using a specified reader function}
\usage{
read.rois(file_paths, rdr = read.polygon_roi, ...)
}
\arguments{
\item{file_paths}{list of file_paths to read}

\item{rdr}{reader function that takes a file path as input and returns
an owin object.}

\item{...}{passed to reader function}
}
\value{
\code{list(\link{owin})}
}
\description{
Function to read multiple windows using a specified reader function
}

