% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{start_matlab}
\alias{start_matlab}
\title{Utility function to start matlab server using R.matlab.}
\usage{
start_matlab(matlab_path = NULL, port = 9999, remote = FALSE,
  interval = NULL, maxTries = NULL)
}
\arguments{
\item{matlab_path}{path to the matlab executable}

\item{port}{port to start matlab server on}

\item{remote}{If TRUE, all data to and from the MATLAB server will be
transferred through the socket connection, otherwise the data will be
transferred via a temporary file.}

\item{interval}{interval at which to poll server to check for results.}

\item{maxTries}{maximum number of times to poll server for results.}
}
\value{
\code{Matlab} instance
}
\description{
Simple wrapper function for starting the matlab server using R.matlab
}

