% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{watershed}
\alias{watershed}
\title{Watershed segmentation of spatstat im object. Requires imager package
\link{\code{watershed}} function to perform watershed segmentation.}
\usage{
watershed(im, seeds, fill_lines = TRUE)
}
\arguments{
\item{im}{image object (grayscale or binary)}

\item{seeds}{planar point pattern containing seed/marker coordinates}
}
\value{
label image.
}
\description{
Watershed segmentation of spatstat im object. Requires imager package
\link{\code{watershed}} function to perform watershed segmentation.
}

