% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{with_matlab}
\alias{with_matlab}
\title{Run matlab code using R.matlab}
\usage{
with_matlab(code, ...)
}
\arguments{
\item{code}{code block to be executed}

\item{...}{passed as arguments to \code{\link{start_matlab}}}
}
\value{
results of code block
}
\description{
Utility to create mini-environment that starts a matlab server, runs
some R.matlab code and then shuts down the matlab server. A fully initialised
R.matlab \code{\link{Matlab}} instance connected to a Matlab server will be
substituted into any reference to the variable \code{matlab} in the code
passed to the with_matlab function.
}

