% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{watershed}
\alias{watershed}
\title{Watershed segmentation of spatstat im object.}
\usage{
watershed(mask, seeds, fill_lines = TRUE)
}
\arguments{
\item{mask}{image object (grayscale or binary)}

\item{seeds}{planar point pattern containing seed/marker coordinates}

\item{fill_lines}{Logical value indicating whether watershed lines should be
filled or left as background.}
}
\value{
labelled image.
}
\description{
Basically a wrapper around the imager package
\code{\link[imager]{watershed}} function to perform marker assisted watershed
segmentation of spatstat \code{im} objects.
}

