% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{with_matlab}
\alias{with_matlab}
\title{Run matlab code using R.matlab}
\usage{
with_matlab(fn, trials = 60, interval = 1, ...)
}
\arguments{
\item{fn}{Single arity function that accepts an initialised/connected
\code{\link{Matlab}} object as input.}

\item{...}{passed as arguments to \code{\link{start_matlab}}}
}
\value{
results of the input \code{fn} function
}
\description{
Utility that starts a matlab server, initialises an R.matlab
\code{\link{Matlab}} instance connected to the Matlab server and passes it to
a single arity input function. Therefore code inside the input function has
access to the \code{\link{Matlab}} instance. Once the function has completed
execution or if a failure occurs, the \code{\link{Matlab}} instance is
disconnected and the Matlab server shutdown.
}

